backend/requirements.txt
flask
Flask-SQLAlchemy
python-dotenv
Flask-Migrate
flask_bcrypt
flask_jwt_extended
flask_cors